Pol 934

+++ source/CsQuery/Dom/Implementation/HtmlElements/FormSubmittableElement.cs
@@ -16,7 +16,7 @@ namespace CsQuery.Implementation
         /// Constructor to specify the element's token ID.
         /// </summary>
         /// <param name="tokenId">The token ID of the element.</param>
-        protected FormSubmittableElement(ushort tokenId)
+        protected FormSubmittableElement(ulong tokenId)
             : base(tokenId)
         {
         }
